+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 0.8.3
+
+*May 21, 2018*
+
+FEATURES:
+
+ - [common] ASCIITrim()
+
+## 0.8.2 (April 23rd, 2018)
+
+FEATURES:
+
+ - [pubsub] TagMap, NewTagMap
+ - [merkle] SimpleProofsFromMap()
+ - [common] IsASCIIText()
+ - [common] PrefixEndBytes // e.g. increment or nil
+ - [common] BitArray.MarshalJSON/.UnmarshalJSON
+ - [common] BitArray uses 'x' not 'X' for String() and above.
+ - [db] DebugDB shows better colorized output
+
+BUG FIXES:
+
+ - [common] Fix TestParallelAbort nondeterministic failure #201/#202
+ - [db] PrefixDB Iterator/ReverseIterator fixes
+ - [db] DebugDB fixes
+
 ## 0.8.1 (April 5th, 2018)
 
 FEATURES: